Prefer indentation over aligning
Ie prefer

    @invoices = current_user.
      invoices.
      order("created_at DESC")

over

    @invoices = current_user.
                invoices.
                order("created_at DESC")

When aligned you'd have to move the lines when changing the variable name, which
is pointless. Also, the length of the variable name might cause the indentation
of the lines to not match tab settings, which in return means more busywork
aligning stuff one character at a time.

Aligning stuff is pointless busywork.