Don't worry about methods named find_by
For example

    def find_by_email(email)
      ...
    end

would be a perfectly good use of a finder. It isn't dynamic, it's clear what it
does, and we can do stuff to the input data beforehand.

Had we used `find_by(:email => email)` we'd be hard pressed to make the search
case insensitive on PostgreSQL, for example.