Opt in to new cops as they are added
This prevents future messages like

> The following cops were added to RuboCop, but are not configured.
> Please set Enabled to either `true` or `false` in your `.rubocop.yml`
> file
> ...

and subsequent annoying configs with a ton of `Enabled: true` blocks.

This way we stay as close to Rubocop standard practices as possible
while still leaving us the option of _disabling_ a cop if it turns out
we prefer something else, instead of having to manually enable cops that
we don't know if we like.

For more information: https://docs.rubocop.org/rubocop/versioning.html